+++ src/views/UndockedViewProvider.ts
@@ -1188,6 +1188,12 @@ export class UndockedViewProvider {
             case "sync":
                 await runGitOperationFromPanel(actionDeps, msg.type);
                 break;
+            case "openRepository":
+                await vscode.commands.executeCommand(
+                    "intelligit.openRepository",
+                    this.selectedRepositoryRoot,
+                );
+                break;
             case "publishBranch":
                 await publishBranchFromPanel(fileActionDeps);
                 break;